comment automatic update
